# backend/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Recipick API
  version: 1.0.0
servers:
  - url: https://api.example.com/v1
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
            details: {}
    RecipeTag:
      type: object
      required: [name, tagType]
      properties:
        name:
          type: string
        tagType:
          type: string
          enum: [system, user]
    Recipe:
      type: object
      required: [id, sourceUrl, title, ingredients, tags, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        sourceUrl:
          type: string
          format: uri
        title:
          type: string
        note:
          type: string
          nullable: true
        servings:
          type: string
          nullable: true
        totalMinutes:
          type: integer
          nullable: true
        ingredients:
          type: array
          items:
            type: string
        tags:
          type: array
          items:
            $ref: '#/components/schemas/RecipeTag'
        sourceRecipeJsonLd:
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ImportPreview:
      type: object
      required: [sourceUrl, ingredients, systemTags, jsonLdFound, warnings]
      properties:
        sourceUrl:
          type: string
          format: uri
        title:
          type: string
          nullable: true
        servings:
          type: string
          nullable: true
        totalMinutes:
          type: integer
          nullable: true
        ingredients:
          type: array
          items:
            type: string
        systemTags:
          type: array
          items:
            type: string
        rawJsonLd:
          nullable: true
        jsonLdFound:
          type: boolean
        warnings:
          type: array
          items:
            type: string
paths:
  /recipes/import:
    post:
      summary: Import recipe preview from URL
      description: Returns preview only. Never saves to DB. Returns 200 even when JSON-LD parse fails.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [sourceUrl]
              properties:
                sourceUrl:
                  type: string
                  format: uri
      responses:
        '200':
          description: Import preview
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImportPreview'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /recipes:
    post:
      summary: Create recipe
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [sourceUrl, title]
              properties:
                sourceUrl:
                  type: string
                  format: uri
                title:
                  type: string
                note:
                  type: string
                  nullable: true
                servings:
                  type: string
                  nullable: true
                totalMinutes:
                  type: integer
                  nullable: true
                ingredients:
                  type: array
                  items:
                    type: string
                tags:
                  type: array
                  items:
                    $ref: '#/components/schemas/RecipeTag'
                sourceRecipeJsonLd:
                  nullable: true
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Recipe'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Duplicate URL
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    get:
      summary: Search recipes
      parameters:
        - in: query
          name: q
          schema:
            type: string
        - in: query
          name: tags
          description: comma separated
          schema:
            type: string
        - in: query
          name: ingredients
          description: comma separated
          schema:
            type: string
        - in: query
          name: maxMinutes
          schema:
            type: integer
            minimum: 0
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 100
        - in: query
          name: offset
          schema:
            type: integer
            minimum: 0
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Recipe'
        '400':
          description: Invalid query
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /recipes/{id}:
    get:
      summary: Get recipe by id
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Recipe'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    patch:
      summary: Update recipe
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                sourceUrl:
                  type: string
                  format: uri
                title:
                  type: string
                note:
                  type: string
                  nullable: true
                servings:
                  type: string
                  nullable: true
                totalMinutes:
                  type: integer
                  nullable: true
                ingredients:
                  type: array
                  items:
                    type: string
                tags:
                  type: array
                  items:
                    $ref: '#/components/schemas/RecipeTag'
                sourceRecipeJsonLd:
                  nullable: true
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Recipe'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Duplicate URL
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      summary: Delete recipe
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Deleted
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'